Moved recommendations logic to helpers
