+++ projects/views.py
@@ -1,5 +1,3 @@
-from random import sample
-
 from django.contrib.auth import get_user_model
 from django.db.models import Q
 from django_filters import rest_framework as filters
@@ -10,7 +8,8 @@ from rest_framework.views import APIView
 
 from core.permissions import IsStaffOrReadOnly
 from projects.filters import ProjectFilter
-from projects.constants import VERBOSE_STEPS, RECOMMENDATIONS_COUNT
+from projects.constants import VERBOSE_STEPS
+from projects.helpers import get_recommended_users
 from projects.models import Project, Achievement
 from projects.permissions import (
     IsProjectLeaderOrReadOnlyForNonDrafts,
@@ -112,27 +111,8 @@ class ProjectRecommendedUsers(generics.RetrieveAPIView):
 
     def get(self, request, pk, **kwargs):
         project = self.get_object()
-
-        # fixme: store key_skills and required_skills more convenient, not just as a string
-        all_needed_skills = set()
-        for vacancy in project.vacancies.all():
-            all_needed_skills.update(set(vacancy.required_skills.lower().split(",")))
-
-        recommended_users = []
-        for user in User.objects.get_members():
-            if user == request.user or not user.key_skills:
-                continue
-
-            skills = set(user.key_skills.lower().split(","))
-            if skills.intersection(all_needed_skills):
-                recommended_users.append(user)
-
-        # get some random users
-        sampled_recommended_users = sample(
-            recommended_users, min(RECOMMENDATIONS_COUNT, len(recommended_users))
-        )
-
-        serializer = self.get_serializer(sampled_recommended_users, many=True)
+        recommended_users = get_recommended_users(project)
+        serializer = self.get_serializer(recommended_users, many=True)
         return Response(status=status.HTTP_200_OK, data=serializer.data)
